Tags for functions include full signature.
Instead of searching for "foo" the user will have to search for "foo()"
or "foo(int)".†  Emacs, at least, offers tab-completion on available
tag names so I don't think I will provide separate "foo" and "foo()"
tags.

† Unless the editor can match a partially-matching tag -- Emacs doesn't
do this on the explicit tagname, only on the pattern field. (The tagname
is the second field, the pattern field the first, which in often just
the whole source line. See etc/ETAGS.EBNF in the Emacs source tree for
an explanation of the TAGS file format.)