Sanity checks for test cases with only "assert_no_tag" assertions.
I don't know why the "set -e" (at run-tests.sh line "(set -e; PATH=... $t)")
isn't having an effect; but a failure at the beginning of the test (e.g. due
to a mis-spelling of the clang-ctags program) doesn't cause the test to stop
immediately, and since the TAGS file is truncated by the ">", the
assert_no_tag would succeed.