Check for clang diagnostics.
A C++ source file with obvious errors wasn't flagged by index.parse, and
I could still iterate over all the cursors. The type of "booyah i;" was
happily reported as TypeKind.INT.

So now I fail to proceed if clang reported any diagnostics.

Note that providing argv[0] (clang-ctag's own program name) to
index.parse caused the following diagnostic on every invocation:

    ./clang-ctags: 'linker' input unused when '-fsyntax-only' is present

In other words, "clang-ctags" was being treated as an additional
argument to the compiler, and since it didn't end in a recognised
extension like ".cpp", it wasn't treated as the source file to compile
but as input to the linker.

If there is a compiler name at argv[0] (instead of "clang-ctags") the
diagnostic is the same; clearly clang_parseTranslationUnit does not
expect to see the name of the compiler itself on the compilation command
line.

Changing argv to argv[1:] eliminates the diagnostic.