Nicer error message when unable to process command line.
The test added in this commit doesn't actually test the specific output
in this error condition, so the test would also pass before this commit.
It is there to test that clang-ctags will exit with a failure status.

When I add command-line flags for clang-ctags (as opposed to the flags
in the compiler command line, passed on to clang) I will have a way of
switching debug on and off -- for now I'll just always print it.