Cleanup: Move clang processing from main into do_tags function
Essentially all I've done is move the middle section of main() into its
own function. There are no functional changes here. This is in
preparation for the upcoming work on supporting a database of compile
commands.

The changes to the Etags and Ctags formatter interface were necessary
because now I instantiate the formatter before clang processing, i.e.
before I know the filename (tu.spelling). This is also a useful for the
upcoming compile-commands work, where the formatter will be able to
generate tags for multiple source files.