Basic support for database of compilation commands
Instead of passing the compiler command line as arguments to
clang-ctags, you can have a json database with the compilation command
line for each source file in your project. If your project uses the
CMake build system, you can generate such a database with
`cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=1`.

So instead of saying:
    clang-ctags -- -Isomewhere -Dsomething -c subdir/a.cpp
you say:
    clang-ctags --compile-commands=path/to/compile_commands.json subdir/a.cpp

Note that `getCompileCommands(source_file)` can return None (if the
source_file isn't present in the database). The user of clang-ctags
could pass in a list of files like "foo.cpp foo.h" where foo.cpp is
present in the database, but foo.h isn't, so it is not an error if the
source file isn't found in the database. In a future commit I will
also generate tags for "foo.h" if it is specified on the command line.

Note that there may be multiple compilation command lines for a single
source file: The same source file could be re-used (say, with different
-D pre-processor defines) to produce different object files. So I
produce tags for *all* compilations of that source file; in a future
commit I'll make the Etags and Ctags formatters smart enough to remove
duplicate tags.