Process multiple source files (including header files)
specified on command line, when `--compile-commands` is given with the
database for looking up each file's compilation command line.

Note that I can now index header files specified on the command line, as
long as I also specify (on the command line) a cpp file (with an entry
in the compile commands database) that includes the header file.

Up until this commit I had no way of choosing which header files to
index; my choices were to generate tags for:

1. all header files (including every header recursively included by
other headers), or
2. perhaps only header files in the same directory as the source
file (or under the directory where clang-ctags was invoked), or
3. no header files at all.

Number 3 is what I was doing up until this commit, and is still what I
do when invoked without `--compile-commands`: I only generate tags when
cursor.location.file.name == tu.spelling.