Use absolute paths to the source files
when (a) searching for a file in the compile commands database, and
(b) recording the source filename in the tags file.

(a) is because the compile_commands.json file produced by CMake (as of
version 2.8.8, at least) specifies files with absolute paths, and the
libclang routine to retrieve the CompileCommands for a file seems to
search on plain string matches: libclang doesn't try to match relative
pathnames given to clang_CompilationDatabase_getCompileCommands against
absolute pathnames in compile_commands.json.

Currently CMake is the only producer of compile commands databases that
I know of, but for compatibility with possible future changes to CMake
and with other producers of the compile commands database, I should
really try to support relative paths in the database's "file" field.
The best way will be to fix libclang, rather than work around it here.

(b) is to keep things simple (or you can call it lazyness). Ideally the
tags file would comply with the following specification from the
etags(1) man page:

    Files specified with relative file names will be recorded in the tag
    table with file names relative to the directory where the tag table
    resides. If the tag table is in /dev or is the standard output,
    however, the file names are made relative to the working directory.
    Files specified with absolute file names will be recorded with
    absolute file names.

The (future) implementation has a slight complication:

* When --compile-commands isn't specified, we can use tu.spelling to
  determine whether the file was specified on the command line with an
  absolute path.

* When --compile-commands *is* specified, we can't use tu.spelling,
  because that comes from the compile command in the database, not from
  the filename specified on the clang-ctags command line. So we'd have
  to pass the original filename all the way through to Etags.tag, but
  also give Etags.tag the absolute filename (though perhaps Etags.tag
  can determine the absolute filename from cwd + cursor.location.file;
  cursor.location.file includes the pathname as given on the compilation
  command line).