`--all-headers` to generate tags for every header file encountered
This will generate tags for system headers too, so the tags file is
going to be huge; only use it to create a tags file for a single source
file.

I've added this option because generating tag for header files is a bit
awkward with clang-ctags. With GNU etags or exuberant ctags you'd just
say `etags file.c file.h other-file.h` and it will index each file.

clang-ctags needs a compilation command-line, and there is no such
command line for header files (unless you manufacture one based on the
command line for one of your source files). So I've added this tag so
that if you index a single source file, you'll be able to find any
definitions in the included header files.

Unfortunately the definitions of some things will be in other
source files that haven't been indexed at all, so this is only somewhat
useful.

It could turn out to be useful as a tool for delving into your system
headers and learning more about your system!