`--non-system-headers` to generate tags for every non-system header file
I can't seem to tell from a cursor or cursor.location or
cursor.location.file whether a file is a system header or not.

System headers have absolute paths... but so do local headers when
the file name is given as an absolute path.

I can't find anything in the libclang C API either; the accessors of the
python cursor.location.file (.name, .time) are clang_getFileName and
clang_getFileTime in clang/tools/libclang/CIndex.cpp, and there is
nothing else in there. Perhaps I could add it in a way similar to
clang_isFileMultipleIncludeGuarded: Use the HeaderSearch object to
determine if it is a system include. Unlikely to be possible without
employing some heuristic though, so I might as well do that heuristic in
my python code.

Given the above, "non-system-headers" means files with a relative path,
or under the directory where clang-ctags was invoked from.