Debug logging with performance info
The numbers clearly show that most of the time is spent in the parsing
done by libclang itself.

Also print the number of the current file being processed and the total
number of files, because running clang-ctags on a large number of files
takes a *long* time, so it is useful to get a rough progress indication.