`--suppress-qualifier-tags` for each namespace qualifier
For a source file containing:

    namespace ns { class cls { int member; }; }

I generate several tags for `member`:

* ::ns::cls::member
* ns::cls::member
* cls::member
* member

See `test_emacs_understands_nested_scopes` in test/why.sh for a thorough
explanation of the reasons.

I now provide the option to suppress this behaviour, because it isn't
necessary if your editor is smart enough to do a substring match on the
tag name (perhaps even a substring match that knows about C++ namespace
separators).

With Emacs I use "ido-ubiquitous"†, a package that provides the "ido"‡
fuzzy matching everywhere in Emacs, including at the "find-tag" prompt.

† https://github.com/technomancy/ido-ubiquitous
‡ http://emacswiki.org/emacs/InteractivelyDoThings