Performance optimization: Don't parse function bodies
Enabling PARSE_SKIP_FUNCTION_BODIES cuts the total runtime down to 40%.
It also starts reporting `cursor.is_definition()` as *False* for
function definitions, so I have to explicitly check for those in my own
`is_definition` function. I don't know if this change in behaviour is
intentional or a bug.

Unfortunately this means we're probably going to start generating tags
for function declarations as well as definitions, but I'll pay that
price for the performance increase.