--suppress-qualifier-tags is now the default behaviour
Use --extra-qualifier-tags to re-enable this behaviour. The option
--suppress-qualifier-tags remains, but has no effect and will be removed
in a future release.

With the old default, the generated tag files were large enough (60MB+)
to cause my editor (Emacs) performance troubles with fuzzy matching
enabled; and for the user the sheer number of matching tags (in the
tab-completion list presented by Emacs) was often overwhelming.